+++ examples/next.js/apollo.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-	client: {
-		service: { name: 'server', localSchemaFile: './schema.json' },
-		includes: ['./{lib,pages,components}/**/*.{graphql,ts,tsx}'],
-	},
-}