+++ src/jest/server.ts
@@ -1,6 +1,0 @@
-import { rest, graphql } from 'msw'
-import { setupServer } from 'msw/node'
-
-const server = setupServer()
-
-export { server, rest, graphql }