+++ src/util/types.ts
@@ -1,5 +1,6 @@
-interface Headers {
+export interface Headers {
 	get(name: string): string | null
+	forEach(callbackfn: (value: string, key: string) => void): void
 }
 
 export interface RequestResult {
@@ -7,4 +8,10 @@ export interface RequestResult {
 	headers: Headers
 	json: () => Promise<any>
 	status: number
+}
+
+export interface FetchOptions {
+	method?: string
+	headers?: any
+	body?: any
 }